move design component out
