+++ src/frontend/src/locales/de/messages.ts
@@ -1,3 +1,3 @@
 /*eslint-disable*/ export const messages = JSON.parse(
-  '{"MHrjPM":"Name","4GKuCs":"Login fehlgeschlagen","jCsNQS":"Überprüfen Sie Ihre Eingabe und versuchen Sie es erneut.","zM9Wd+":"Anmeldung erfolgreich","rxWA39":"Willkommen zurück!","yfblq9":"Mail erfolgreich gesendet","R2JMfc":"Prüfen Sie Ihren Posteingang auf den Anmeldelink. Wenn Sie ein Konto haben, erhalten Sie einen Anmeldelink. Prüfen Sie auch den Spam.","ccnxuA":"Eingabefehler","7sNhEz":"Nutzername","8ZsakT":"Passwort","9TO8nT":"Dein Passwort","RfwZxd":"Passwort zurücksetzen","O3oNi5":"Mail","Wr5sDQ":"Wir werden Ihnen einen Link für die Anmeldung senden","7ZOmjI":"Mail erhalten","XlWstl":"Benutzername und Passwort verwenden","sQia9P":"Einoggen","F+gz9Z":"Mail senden","Ai2U7L":"Adresse","6YtxFj":"Name","yWMzcH":"Kein Eintrag...","UYWLpE":"Adresse hinzufügen","tfDRzk":"Speichern","IvkbIT":"Mehr lesen","29VNqC":"Unbekannter Fehler","nlJhkA":"Ein Fehler ist aufgetreten:","DVAy0b":"InvenTree\'s Logo","3TnJRX":"This feature/button/site is a placeholder for a feature that is not implemented, only partial or intended for testing.","etqXdW":"PLH","XDwkfO":"Scan QR code","GU7xAr":"Unknown response","UHot+L":"Error while getting camera","bR26mb":"Error while scanning","fvJQqd":"Error while stopping","CMQ09J":"Scanning","Fg9r/3":"Not scanning","QuNKRX":"Select Camera","m3BKG+":"Start scanning","yFRXH8":"Stop scanning","3164SS":"No scans yet!","RWw9Lg":"Dialog schließen","iDNBZe":"Benachrichtigungen","vERlcd":"Profil","Tz0i8g":"Einstellungen","r5Xdbs":["Aktuelle Sprache ",["locale"]],"XXvCbv":"Zu Pseudo-Sprache wechseln","T3FM0r":"Benutzereinstellungen","nOhz3x":"Abmelden","rmlxV1":"Open Navigation","N6Pxr9":"Alle anzeigen","ZDIydz":"Loslegen","BQDL+H":"Übersicht über die wichtigsten Objekte, Funktionen und mögliche Anwendungsfälle.","UxKoFf":"Navigation","wRR604":"Pages","ohUJJM":"Plugins","TvY/XA":"Documentation","uyJsf6":"About","hQRttt":"Speichern","dEgA5A":"Abbrechen","vgP+9p":"Teil","tMMrz4":"Demo von InvenTree","vu8/DU":"Lokaler Server","On0aF2":"Webseite","RkXlPZ":"GitHub","kc+zZA":"Demo","i0qMbr":"Startseite","4GLxhy":"Getting Started","VAYCzI":"Getting started with InvenTree","OZtEcz":"API","aW0h/b":"InvenTree API documentation","BOAupq":"Developer Manual","kUcL4g":"InvenTree developer manual","/lDBHm":"FAQ","a3pVqb":"Frequently asked questions","kyAi7k":"Instance","Q5S3DY":"About this Inventree instance","vHeNia":"InvenTree","gfhzPz":"About the InvenTree org","snyG+w":"Licenses","tBjIo1":"Licenses for packages used by InvenTree","2AZart":"Open sourcea","v+Wp++":"Open source","fu2+tK":"Start page of your instance.","pTE4nz":"This Pokémon’s cry is very loud and distracting","S+oekQ":"This Pokémon’s cry is very loud and distracting and more and more and more","kNyJAF":"Profile page","CFYxhi":"User attributes and design settings.","uP4V6I":"Free for everyone","dyMOjI":"The fluid of Smeargle’s tail secretions changes","d1WpzX":"The fluid of Smeargle’s tail secretions changes in the intensity","1ekmeV":"abc","Wj+wQW":"Random image","sGi2sH":"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Name liber tempor cum soluta nobis eleifend option congue nihil imperdiet doming id quod mazim placerat facer possim assume. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, At accusam aliquyam diam diam dolore dolores duo eirmod eos erat, et nonumy sed tempor et et invidunt justo labore Stet clita ea et gubergren, kasd magna no rebum. sanctus sea sed takimata ut vero voluptua. est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat. Consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor","K+7Exx":"Yanma is capable of seeing 360 degrees without","a3LDKx":"Security","D9kxcs":"The shell’s rounded shape and the grooves on its.","ZlwDi6":"Analytics","Ntb/Ja":"This Pokémon uses its flying ability to quickly chase","q+Lv8f":"Combusken battles with the intensely hot flames it spews","FKQcYZ":"Erfolgreich abgemeldet","aJhI/3":"Auf Wiedersehen.","eX0txO":"Prüfen Sie Ihren Posteingang für einen Link zum Zurücksetzen. Dies funktioniert nur, wenn Sie ein Konto haben. Prüfen Sie auch den Spam-Ordner.","WhimMi":"Zurücksetzen fehlgeschlagen","iVj6ge":"Bereits angemeldet","FR/+0K":"Found an existing login - using it to log you in.","ps9k8Y":"Prüfe ob Sie bereits angemeldet sind","bX1aQ5":"Keine Auswahl","AA2j+t":"Server konfigurieren","eV2FZ+":"Token ungültig","uAHzZQ":"Sie müssen einen gültigen Token angeben, um ein neues Passwort festzulegen. Prüfen Sie Ihren Posteingang für einen Link zum Zurücksetzen.","+5xxir":"Kein Token angegeben","KuLTFa":"Sie müssen einen Token angeben, um ein neues Passwort festzulegen. Prüfen Sie Ihren Posteingang für einen Link zum Zurücksetzen.","Hw2MHB":"Passwort festgelegt","+p8fKY":"Das Passwort wurde erfolgreich festgelegt. Sie können sich jetzt mit Ihrem neuen Passwort anmelden","V/e7nf":"Passwort festlegen","SlfejT":"Fehler","TpqeIh":["Fehler: ",["0"]],"b3ilvM":"Es ist ein unerwarteter Fehler aufgetreten.","5QTyaY":"Subscribed Parts","GuGbPw":"Subscribed Categories","LcKNFQ":"Latest Parts","eHUZsJ":"BOM Waiting Validation","ZopSbj":"Recently Updated","UgdO7s":"Low Stock","Onj2Pw":"Depleted Stock","Iq/utX":"Required for Build Orders","ZOsmSm":"Expired Stock","kc9cAF":"Stale Stock","zLhIiS":"Build Orders In Progress","UBWkDy":"Overdue Build Orders","WsHr9R":"Outstanding Purchase Orders","fCNzWA":"Overdue Purchase Orders","gyZThB":"Outstanding Sales Orders","Gu8K8T":"Overdue Sales Orders","XzTq3p":"Current News","7p5kLi":"Übersicht","edpMcF":"Automatisch aktualisieren","7PzzBU":"Nutzer","ekfzWq":"Benutzereinstellungen","hFwWnI":"Benachrichtigungseinstellungen","50nnEk":"Servereinstellungen","eKHY3W":"Plugineinstellungen","c6Mp+A":"Einstellungen für den aktuellen Benutzer","EBBDLp":"Home Page Settings","d42r7C":"Search Settings","o0PqeM":"Etikettendruck","nutMuO":"Berichte","DdjH42":"Display Settings","VzYWwh":"Settings for the notifications","0fzps+":"Global Server Settings","5u2+so":"Global Settings for this instance","S60KP9":"Serverkonfiguration","R+R5Sa":"Anmeldeeinstellungen","NP6Hng":"Barcode-Feature verwenden","H2tPtY":"Teile","axC9dx":"Bepreisung","PN5rCS":"Bestand","1PGWAQ":"Bauaufträge","WObPen":"Bestellugnen","H7F6Gx":"Aufträge","Ud411M":"Plugin Settings for this instance","pkdjGY":"Data is current beeing loaded","gIQQwD":"Failed to load","UbtqIw":"Show internal names","UIwUzc":["Input ",["0"]," is not known"],"J9kB0C":["Saved changes ",["0"]],"EOyF2I":["Changed to ",["0"]],"pa6s4O":["Error while saving ",["0"]],"/JfytP":["Error was ",["err"]],"IBGfrY":["Plugin: ",["0"]],"Se2ost":["Method: ",["0"]],"nDqlBl":"Benutzerdetails","JOUEkZ":["Vorname: ",["0"]],"GlGzeI":["Nachname: ",["0"]],"PkcDO7":["Benutzername: ",["0"]],"xyAcm2":"Eine bestehende Anmeldung wurde gefunden - diese wird verwendet um Sie anzumelden.","pAtylB":"Nicht gefunden","FeQ++0":"Diese Seite ist nicht bekannt oder wurde verschoben.","wmCIch":"Zur Startseite","OWg6Ht":"Mobile viewport detected","j1oKmM":"Platform UI is optimized for Tablets and Desktops, you can use the official app for a mobile experience.","NtcqDr":"Read the docs"}'
+  '{"MHrjPM":"Name","4GKuCs":"Login fehlgeschlagen","jCsNQS":"Überprüfen Sie Ihre Eingabe und versuchen Sie es erneut.","zM9Wd+":"Anmeldung erfolgreich","rxWA39":"Willkommen zurück!","yfblq9":"Mail erfolgreich gesendet","R2JMfc":"Prüfen Sie Ihren Posteingang auf den Anmeldelink. Wenn Sie ein Konto haben, erhalten Sie einen Anmeldelink. Prüfen Sie auch den Spam.","ccnxuA":"Eingabefehler","7sNhEz":"Nutzername","8ZsakT":"Passwort","9TO8nT":"Dein Passwort","RfwZxd":"Passwort zurücksetzen","O3oNi5":"Mail","Wr5sDQ":"Wir werden Ihnen einen Link für die Anmeldung senden","7ZOmjI":"Mail erhalten","XlWstl":"Benutzername und Passwort verwenden","sQia9P":"Einoggen","F+gz9Z":"Mail senden","Ai2U7L":"Adresse","6YtxFj":"Name","yWMzcH":"Kein Eintrag...","UYWLpE":"Adresse hinzufügen","tfDRzk":"Speichern","IvkbIT":"Mehr lesen","29VNqC":"Unbekannter Fehler","nlJhkA":"Ein Fehler ist aufgetreten:","DVAy0b":"InvenTree\'s Logo","3TnJRX":"This feature/button/site is a placeholder for a feature that is not implemented, only partial or intended for testing.","etqXdW":"PLH","XDwkfO":"Scan QR code","GU7xAr":"Unknown response","UHot+L":"Error while getting camera","bR26mb":"Error while scanning","fvJQqd":"Error while stopping","CMQ09J":"Scanning","Fg9r/3":"Not scanning","QuNKRX":"Select Camera","m3BKG+":"Start scanning","yFRXH8":"Stop scanning","3164SS":"No scans yet!","RWw9Lg":"Dialog schließen","iDNBZe":"Benachrichtigungen","vERlcd":"Profil","Tz0i8g":"Einstellungen","r5Xdbs":["Aktuelle Sprache ",["locale"]],"XXvCbv":"Zu Pseudo-Sprache wechseln","T3FM0r":"Benutzereinstellungen","nOhz3x":"Abmelden","rmlxV1":"Open Navigation","N6Pxr9":"Alle anzeigen","ZDIydz":"Loslegen","BQDL+H":"Übersicht über die wichtigsten Objekte, Funktionen und mögliche Anwendungsfälle.","UxKoFf":"Navigation","wRR604":"Pages","ohUJJM":"Plugins","TvY/XA":"Documentation","uyJsf6":"About","hQRttt":"Speichern","dEgA5A":"Abbrechen","vgP+9p":"Teil","tMMrz4":"Demo von InvenTree","vu8/DU":"Lokaler Server","On0aF2":"Webseite","RkXlPZ":"GitHub","kc+zZA":"Demo","i0qMbr":"Startseite","4GLxhy":"Getting Started","VAYCzI":"Getting started with InvenTree","OZtEcz":"API","aW0h/b":"InvenTree API documentation","BOAupq":"Developer Manual","kUcL4g":"InvenTree developer manual","/lDBHm":"FAQ","a3pVqb":"Frequently asked questions","kyAi7k":"Instance","Q5S3DY":"About this Inventree instance","vHeNia":"InvenTree","gfhzPz":"About the InvenTree org","snyG+w":"Licenses","tBjIo1":"Licenses for packages used by InvenTree","2AZart":"Open sourcea","v+Wp++":"Open source","fu2+tK":"Start page of your instance.","pTE4nz":"This Pokémon’s cry is very loud and distracting","S+oekQ":"This Pokémon’s cry is very loud and distracting and more and more and more","kNyJAF":"Profile page","CFYxhi":"User attributes and design settings.","uP4V6I":"Free for everyone","dyMOjI":"The fluid of Smeargle’s tail secretions changes","d1WpzX":"The fluid of Smeargle’s tail secretions changes in the intensity","1ekmeV":"abc","Wj+wQW":"Random image","sGi2sH":"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Name liber tempor cum soluta nobis eleifend option congue nihil imperdiet doming id quod mazim placerat facer possim assume. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, At accusam aliquyam diam diam dolore dolores duo eirmod eos erat, et nonumy sed tempor et et invidunt justo labore Stet clita ea et gubergren, kasd magna no rebum. sanctus sea sed takimata ut vero voluptua. est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat. Consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor","K+7Exx":"Yanma is capable of seeing 360 degrees without","a3LDKx":"Security","D9kxcs":"The shell’s rounded shape and the grooves on its.","ZlwDi6":"Analytics","Ntb/Ja":"This Pokémon uses its flying ability to quickly chase","q+Lv8f":"Combusken battles with the intensely hot flames it spews","FKQcYZ":"Erfolgreich abgemeldet","aJhI/3":"Auf Wiedersehen.","eX0txO":"Prüfen Sie Ihren Posteingang für einen Link zum Zurücksetzen. Dies funktioniert nur, wenn Sie ein Konto haben. Prüfen Sie auch den Spam-Ordner.","WhimMi":"Zurücksetzen fehlgeschlagen","iVj6ge":"Bereits angemeldet","FR/+0K":"Found an existing login - using it to log you in.","ps9k8Y":"Prüfe ob Sie bereits angemeldet sind","bX1aQ5":"Keine Auswahl","AA2j+t":"Server konfigurieren","eV2FZ+":"Token ungültig","uAHzZQ":"Sie müssen einen gültigen Token angeben, um ein neues Passwort festzulegen. Prüfen Sie Ihren Posteingang für einen Link zum Zurücksetzen.","+5xxir":"Kein Token angegeben","KuLTFa":"Sie müssen einen Token angeben, um ein neues Passwort festzulegen. Prüfen Sie Ihren Posteingang für einen Link zum Zurücksetzen.","Hw2MHB":"Passwort festgelegt","+p8fKY":"Das Passwort wurde erfolgreich festgelegt. Sie können sich jetzt mit Ihrem neuen Passwort anmelden","V/e7nf":"Passwort festlegen","SlfejT":"Fehler","TpqeIh":["Fehler: ",["0"]],"b3ilvM":"Es ist ein unerwarteter Fehler aufgetreten.","5QTyaY":"Subscribed Parts","GuGbPw":"Subscribed Categories","LcKNFQ":"Latest Parts","eHUZsJ":"BOM Waiting Validation","ZopSbj":"Recently Updated","UgdO7s":"Low Stock","Onj2Pw":"Depleted Stock","Iq/utX":"Required for Build Orders","ZOsmSm":"Expired Stock","kc9cAF":"Stale Stock","zLhIiS":"Build Orders In Progress","UBWkDy":"Overdue Build Orders","WsHr9R":"Outstanding Purchase Orders","fCNzWA":"Overdue Purchase Orders","gyZThB":"Outstanding Sales Orders","Gu8K8T":"Overdue Sales Orders","XzTq3p":"Current News","7p5kLi":"Übersicht","edpMcF":"Automatisch aktualisieren","7PzzBU":"Nutzer","ekfzWq":"Benutzereinstellungen","hFwWnI":"Benachrichtigungseinstellungen","50nnEk":"Servereinstellungen","eKHY3W":"Plugineinstellungen","c6Mp+A":"Einstellungen für den aktuellen Benutzer","EBBDLp":"Home Page Settings","d42r7C":"Search Settings","o0PqeM":"Etikettendruck","nutMuO":"Berichte","DdjH42":"Display Settings","VzYWwh":"Settings for the notifications","0fzps+":"Global Server Settings","5u2+so":"Global Settings for this instance","S60KP9":"Serverkonfiguration","R+R5Sa":"Anmeldeeinstellungen","NP6Hng":"Barcode-Feature verwenden","H2tPtY":"Teile","axC9dx":"Bepreisung","PN5rCS":"Bestand","1PGWAQ":"Bauaufträge","WObPen":"Bestellugnen","H7F6Gx":"Aufträge","Ud411M":"Plugin Settings for this instance","pkdjGY":"Data is current beeing loaded","gIQQwD":"Failed to load","UbtqIw":"Show internal names","UIwUzc":["Input ",["0"]," is not known"],"J9kB0C":["Saved changes ",["0"]],"EOyF2I":["Changed to ",["0"]],"pa6s4O":["Error while saving ",["0"]],"/JfytP":["Error was ",["err"]],"IBGfrY":["Plugin: ",["0"]],"Se2ost":["Method: ",["0"]],"nDqlBl":"Benutzerdetails","JOUEkZ":["Vorname: ",["0"]],"GlGzeI":["Nachname: ",["0"]],"PkcDO7":["Benutzername: ",["0"]],"M46ISI":"bars","Ai6veK":"oval","8zGXnJ":"dots","gDIqhx":"Design <0/>","QFd2P1":"Primary color","160vo+":"White color","u01284":"Black color","bjp1xg":"Border Radius","EBeoY+":"Loader","xyAcm2":"Eine bestehende Anmeldung wurde gefunden - diese wird verwendet um Sie anzumelden.","pAtylB":"Nicht gefunden","FeQ++0":"Diese Seite ist nicht bekannt oder wurde verschoben.","wmCIch":"Zur Startseite","OWg6Ht":"Mobile viewport detected","j1oKmM":"Platform UI is optimized for Tablets and Desktops, you can use the official app for a mobile experience.","NtcqDr":"Read the docs"}'
 );
