+++ src/frontend/src/locales/en/messages.po
@@ -256,7 +256,7 @@ msgid "About"
 msgstr "About"
 
 #: src/contexts/ThemeContext.tsx:62
-#: src/pages/Index/Profile/UserPanel.tsx:132
+#: src/pages/Index/Profile/UserPanel.tsx:121
 msgid "Submit"
 msgstr "Submit"
 
@@ -724,21 +724,57 @@ msgstr "User"
 #~ msgid "Method: {0}"
 #~ msgstr "Method: {0}"
 
-#: src/pages/Index/Profile/UserPanel.tsx:90
+#: src/pages/Index/Profile/UserPanel.tsx:79
 msgid "Userinfo"
 msgstr "Userinfo"
 
-#: src/pages/Index/Profile/UserPanel.tsx:103
+#: src/pages/Index/Profile/UserPanel.tsx:92
 msgid "First name: {0}"
 msgstr "First name: {0}"
 
-#: src/pages/Index/Profile/UserPanel.tsx:114
+#: src/pages/Index/Profile/UserPanel.tsx:103
 msgid "Last name: {0}"
 msgstr "Last name: {0}"
 
-#: src/pages/Index/Profile/UserPanel.tsx:125
+#: src/pages/Index/Profile/UserPanel.tsx:114
 msgid "Username: {0}"
 msgstr "Username: {0}"
+
+#: src/pages/Index/Profile/UserTheme.tsx:71
+msgid "bars"
+msgstr "bars"
+
+#: src/pages/Index/Profile/UserTheme.tsx:72
+msgid "oval"
+msgstr "oval"
+
+#: src/pages/Index/Profile/UserTheme.tsx:73
+msgid "dots"
+msgstr "dots"
+
+#: src/pages/Index/Profile/UserTheme.tsx:84
+msgid "Design <0/>"
+msgstr "Design <0/>"
+
+#: src/pages/Index/Profile/UserTheme.tsx:92
+msgid "Primary color"
+msgstr "Primary color"
+
+#: src/pages/Index/Profile/UserTheme.tsx:105
+msgid "White color"
+msgstr "White color"
+
+#: src/pages/Index/Profile/UserTheme.tsx:113
+msgid "Black color"
+msgstr "Black color"
+
+#: src/pages/Index/Profile/UserTheme.tsx:121
+msgid "Border Radius"
+msgstr "Border Radius"
+
+#: src/pages/Index/Profile/UserTheme.tsx:137
+msgid "Loader"
+msgstr "Loader"
 
 #: src/pages/Logged-In.tsx:24
 #~ msgid "Found an exsisting login - using it to log you in."
