+++ src/frontend/src/locales/hu/messages.ts
@@ -1,3 +1,3 @@
 /*eslint-disable*/ export const messages = JSON.parse(
-  '{"MHrjPM":"Title","4GKuCs":"Login failed","jCsNQS":"Check your your input and try again.","zM9Wd+":"Login successfull","rxWA39":"Welcome back!","yfblq9":"Mail delivery successfull","R2JMfc":"Check your inbox for the login link. If you have an account, you will receive a login link. Check in spam too.","ccnxuA":"Input error","7sNhEz":"Username","8ZsakT":"Password","9TO8nT":"Your password","RfwZxd":"Reset password","O3oNi5":"Email","Wr5sDQ":"We will send you a link to login - if you are registered","7ZOmjI":"Send me an email","XlWstl":"I will use username and password","sQia9P":"Log in","F+gz9Z":"Send mail","Ai2U7L":"Host","6YtxFj":"Name","yWMzcH":"No one here...","UYWLpE":"Add Host","tfDRzk":"Save","IvkbIT":"Read More","29VNqC":"Unknown error","nlJhkA":"An error occurred:","DVAy0b":"InvenTree Logo","3TnJRX":"This feature/button/site is a placeholder for a feature that is not implemented, only partial or intended for testing.","etqXdW":"PLH","XDwkfO":"Scan QR code","GU7xAr":"Unknown response","UHot+L":"Error while getting camera","bR26mb":"Error while scanning","fvJQqd":"Error while stopping","CMQ09J":"Scanning","Fg9r/3":"Not scanning","QuNKRX":"Select Camera","m3BKG+":"Start scanning","yFRXH8":"Stop scanning","3164SS":"No scans yet!","RWw9Lg":"Close modal","iDNBZe":"Notifications","vERlcd":"Profile","Tz0i8g":"Settings","r5Xdbs":["Current language ",["locale"]],"XXvCbv":"Switch to pseudo language","T3FM0r":"Account settings","nOhz3x":"Logout","rmlxV1":"Open Navigation","N6Pxr9":"View all","ZDIydz":"Get started","BQDL+H":"Overview over high-level objects, functions and possible usecases.","UxKoFf":"Navigation","wRR604":"Pages","ohUJJM":"Plugins","TvY/XA":"Documentation","uyJsf6":"About","hQRttt":"Submit","dEgA5A":"Cancel","vgP+9p":"Alkatrész","tMMrz4":"InvenTree Demo","vu8/DU":"Local Server","On0aF2":"Website","RkXlPZ":"GitHub","kc+zZA":"Demo","i0qMbr":"Home","4GLxhy":"Getting Started","VAYCzI":"Getting started with InvenTree","OZtEcz":"API","aW0h/b":"InvenTree API documentation","BOAupq":"Developer Manual","kUcL4g":"InvenTree developer manual","/lDBHm":"FAQ","a3pVqb":"Frequently asked questions","kyAi7k":"Instance","Q5S3DY":"About this Inventree instance","vHeNia":"InvenTree","gfhzPz":"About the InvenTree org","snyG+w":"Licenses","tBjIo1":"Licenses for packages used by InvenTree","2AZart":"Open sourcea","v+Wp++":"Open source","fu2+tK":"Start page of your instance.","pTE4nz":"This Pokémon’s cry is very loud and distracting","S+oekQ":"This Pokémon’s cry is very loud and distracting and more and more and more","kNyJAF":"Profile page","CFYxhi":"User attributes and design settings.","uP4V6I":"Free for everyone","dyMOjI":"The fluid of Smeargle’s tail secretions changes","d1WpzX":"The fluid of Smeargle’s tail secretions changes in the intensity","1ekmeV":"abc","Wj+wQW":"Random image","sGi2sH":"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Name liber tempor cum soluta nobis eleifend option congue nihil imperdiet doming id quod mazim placerat facer possim assume. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, At accusam aliquyam diam diam dolore dolores duo eirmod eos erat, et nonumy sed tempor et et invidunt justo labore Stet clita ea et gubergren, kasd magna no rebum. sanctus sea sed takimata ut vero voluptua. est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat. Consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor","K+7Exx":"Yanma is capable of seeing 360 degrees without","a3LDKx":"Security","D9kxcs":"The shell’s rounded shape and the grooves on its.","ZlwDi6":"Analytics","Ntb/Ja":"This Pokémon uses its flying ability to quickly chase","q+Lv8f":"Combusken battles with the intensely hot flames it spews","FKQcYZ":"Logout successfull","aJhI/3":"See you soon.","eX0txO":"Check your inbox for a reset link. This only works if you have an account. Check in spam too.","WhimMi":"Reset failed","iVj6ge":"Already logged in","FR/+0K":"Found an existing login - using it to log you in.","ps9k8Y":"Checking if you are already logged in","bX1aQ5":"No selection","AA2j+t":"Edit host options","eV2FZ+":"Token invalid","uAHzZQ":"You need to provide a valid token to set a new password. Check your inbox for a reset link.","+5xxir":"No token provided","KuLTFa":"You need to provide a token to set a new password. Check your inbox for a reset link.","Hw2MHB":"Password set","+p8fKY":"The password was set successfully. You can now login with your new password","V/e7nf":"Set new password","SlfejT":"Error","TpqeIh":["Error: ",["0"]],"b3ilvM":"Sorry, an unexpected error has occurred.","5QTyaY":"Subscribed Parts","GuGbPw":"Subscribed Categories","LcKNFQ":"Latest Parts","eHUZsJ":"BOM Waiting Validation","ZopSbj":"Recently Updated","UgdO7s":"Low Stock","Onj2Pw":"Depleted Stock","Iq/utX":"Required for Build Orders","ZOsmSm":"Expired Stock","kc9cAF":"Stale Stock","zLhIiS":"Build Orders In Progress","UBWkDy":"Overdue Build Orders","WsHr9R":"Outstanding Purchase Orders","fCNzWA":"Overdue Purchase Orders","gyZThB":"Outstanding Sales Orders","Gu8K8T":"Overdue Sales Orders","XzTq3p":"Current News","7p5kLi":"Dashboard","edpMcF":"Autoupdate","7PzzBU":"User","ekfzWq":"User Settings","hFwWnI":"Notification Settings","50nnEk":"Global Settings","eKHY3W":"Plugin Settings","c6Mp+A":"Settings for the current user","EBBDLp":"Home Page Settings","d42r7C":"Search Settings","o0PqeM":"Label Settings","nutMuO":"Report Settings","DdjH42":"Display Settings","VzYWwh":"Settings for the notifications","0fzps+":"Global Server Settings","5u2+so":"Global Settings for this instance","S60KP9":"Server Settings","R+R5Sa":"Login Settings","NP6Hng":"Barcode Settings","H2tPtY":"Part Settings","axC9dx":"Pricing Settings","PN5rCS":"Stock Settings","1PGWAQ":"Build Order Settings","WObPen":"Purchase Order Settings","H7F6Gx":"Sales Order Settings","Ud411M":"Plugin Settings for this instance","pkdjGY":"Data is current beeing loaded","gIQQwD":"Failed to load","UbtqIw":"Show internal names","UIwUzc":["Input ",["0"]," is not known"],"J9kB0C":["Saved changes ",["0"]],"EOyF2I":["Changed to ",["0"]],"pa6s4O":["Error while saving ",["0"]],"/JfytP":["Error was ",["err"]],"IBGfrY":["Plugin: ",["0"]],"Se2ost":["Method: ",["0"]],"nDqlBl":"Userinfo","JOUEkZ":["First name: ",["0"]],"GlGzeI":["Last name: ",["0"]],"PkcDO7":["Username: ",["0"]],"xyAcm2":"Found an exsisting login - using it to log you in.","pAtylB":"Not Found","FeQ++0":"Sorry, this page is not known or was moved.","wmCIch":"Go to the start page","OWg6Ht":"Mobile viewport detected","j1oKmM":"Platform UI is optimized for Tablets and Desktops, you can use the official app for a mobile experience.","NtcqDr":"Read the docs","jdbeGb":"login","Phg8Hk":"register","JrBnCJ":["Welcome ",["actionname"]," to"],"hx1ePY":"Placeholder","g87ZP7":"Or continue with email","q8yluz":"Your name","B2Tpo0":"Invalid email","vz42/T":"Password should include at least 6 characters","3xHmj+":"Already have an account? Login","L5ridO":"Don\'t have an account? Register","oFjpQm":"An error occured:"}'
+  '{"MHrjPM":"Title","4GKuCs":"Login failed","jCsNQS":"Check your your input and try again.","zM9Wd+":"Login successfull","rxWA39":"Welcome back!","yfblq9":"Mail delivery successfull","R2JMfc":"Check your inbox for the login link. If you have an account, you will receive a login link. Check in spam too.","ccnxuA":"Input error","7sNhEz":"Username","8ZsakT":"Password","9TO8nT":"Your password","RfwZxd":"Reset password","O3oNi5":"Email","Wr5sDQ":"We will send you a link to login - if you are registered","7ZOmjI":"Send me an email","XlWstl":"I will use username and password","sQia9P":"Log in","F+gz9Z":"Send mail","Ai2U7L":"Host","6YtxFj":"Name","yWMzcH":"No one here...","UYWLpE":"Add Host","tfDRzk":"Save","IvkbIT":"Read More","29VNqC":"Unknown error","nlJhkA":"An error occurred:","DVAy0b":"InvenTree Logo","3TnJRX":"This feature/button/site is a placeholder for a feature that is not implemented, only partial or intended for testing.","etqXdW":"PLH","XDwkfO":"Scan QR code","GU7xAr":"Unknown response","UHot+L":"Error while getting camera","bR26mb":"Error while scanning","fvJQqd":"Error while stopping","CMQ09J":"Scanning","Fg9r/3":"Not scanning","QuNKRX":"Select Camera","m3BKG+":"Start scanning","yFRXH8":"Stop scanning","3164SS":"No scans yet!","RWw9Lg":"Close modal","iDNBZe":"Notifications","vERlcd":"Profile","Tz0i8g":"Settings","r5Xdbs":["Current language ",["locale"]],"XXvCbv":"Switch to pseudo language","T3FM0r":"Account settings","nOhz3x":"Logout","rmlxV1":"Open Navigation","N6Pxr9":"View all","ZDIydz":"Get started","BQDL+H":"Overview over high-level objects, functions and possible usecases.","UxKoFf":"Navigation","wRR604":"Pages","ohUJJM":"Plugins","TvY/XA":"Documentation","uyJsf6":"About","hQRttt":"Submit","dEgA5A":"Cancel","vgP+9p":"Alkatrész","tMMrz4":"InvenTree Demo","vu8/DU":"Local Server","On0aF2":"Website","RkXlPZ":"GitHub","kc+zZA":"Demo","i0qMbr":"Home","4GLxhy":"Getting Started","VAYCzI":"Getting started with InvenTree","OZtEcz":"API","aW0h/b":"InvenTree API documentation","BOAupq":"Developer Manual","kUcL4g":"InvenTree developer manual","/lDBHm":"FAQ","a3pVqb":"Frequently asked questions","kyAi7k":"Instance","Q5S3DY":"About this Inventree instance","vHeNia":"InvenTree","gfhzPz":"About the InvenTree org","snyG+w":"Licenses","tBjIo1":"Licenses for packages used by InvenTree","2AZart":"Open sourcea","v+Wp++":"Open source","fu2+tK":"Start page of your instance.","pTE4nz":"This Pokémon’s cry is very loud and distracting","S+oekQ":"This Pokémon’s cry is very loud and distracting and more and more and more","kNyJAF":"Profile page","CFYxhi":"User attributes and design settings.","uP4V6I":"Free for everyone","dyMOjI":"The fluid of Smeargle’s tail secretions changes","d1WpzX":"The fluid of Smeargle’s tail secretions changes in the intensity","1ekmeV":"abc","Wj+wQW":"Random image","sGi2sH":"Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore the feugait nulla facilisi. Name liber tempor cum soluta nobis eleifend option congue nihil imperdiet doming id quod mazim placerat facer possim assume. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, At accusam aliquyam diam diam dolore dolores duo eirmod eos erat, et nonumy sed tempor et et invidunt justo labore Stet clita ea et gubergren, kasd magna no rebum. sanctus sea sed takimata ut vero voluptua. est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat. Consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor","K+7Exx":"Yanma is capable of seeing 360 degrees without","a3LDKx":"Security","D9kxcs":"The shell’s rounded shape and the grooves on its.","ZlwDi6":"Analytics","Ntb/Ja":"This Pokémon uses its flying ability to quickly chase","q+Lv8f":"Combusken battles with the intensely hot flames it spews","FKQcYZ":"Logout successfull","aJhI/3":"See you soon.","eX0txO":"Check your inbox for a reset link. This only works if you have an account. Check in spam too.","WhimMi":"Reset failed","iVj6ge":"Already logged in","FR/+0K":"Found an existing login - using it to log you in.","ps9k8Y":"Checking if you are already logged in","bX1aQ5":"No selection","AA2j+t":"Edit host options","eV2FZ+":"Token invalid","uAHzZQ":"You need to provide a valid token to set a new password. Check your inbox for a reset link.","+5xxir":"No token provided","KuLTFa":"You need to provide a token to set a new password. Check your inbox for a reset link.","Hw2MHB":"Password set","+p8fKY":"The password was set successfully. You can now login with your new password","V/e7nf":"Set new password","SlfejT":"Error","TpqeIh":["Error: ",["0"]],"b3ilvM":"Sorry, an unexpected error has occurred.","5QTyaY":"Subscribed Parts","GuGbPw":"Subscribed Categories","LcKNFQ":"Latest Parts","eHUZsJ":"BOM Waiting Validation","ZopSbj":"Recently Updated","UgdO7s":"Low Stock","Onj2Pw":"Depleted Stock","Iq/utX":"Required for Build Orders","ZOsmSm":"Expired Stock","kc9cAF":"Stale Stock","zLhIiS":"Build Orders In Progress","UBWkDy":"Overdue Build Orders","WsHr9R":"Outstanding Purchase Orders","fCNzWA":"Overdue Purchase Orders","gyZThB":"Outstanding Sales Orders","Gu8K8T":"Overdue Sales Orders","XzTq3p":"Current News","7p5kLi":"Dashboard","edpMcF":"Autoupdate","7PzzBU":"User","ekfzWq":"User Settings","hFwWnI":"Notification Settings","50nnEk":"Global Settings","eKHY3W":"Plugin Settings","c6Mp+A":"Settings for the current user","EBBDLp":"Home Page Settings","d42r7C":"Search Settings","o0PqeM":"Label Settings","nutMuO":"Report Settings","DdjH42":"Display Settings","VzYWwh":"Settings for the notifications","0fzps+":"Global Server Settings","5u2+so":"Global Settings for this instance","S60KP9":"Server Settings","R+R5Sa":"Login Settings","NP6Hng":"Barcode Settings","H2tPtY":"Part Settings","axC9dx":"Pricing Settings","PN5rCS":"Stock Settings","1PGWAQ":"Build Order Settings","WObPen":"Purchase Order Settings","H7F6Gx":"Sales Order Settings","Ud411M":"Plugin Settings for this instance","pkdjGY":"Data is current beeing loaded","gIQQwD":"Failed to load","UbtqIw":"Show internal names","UIwUzc":["Input ",["0"]," is not known"],"J9kB0C":["Saved changes ",["0"]],"EOyF2I":["Changed to ",["0"]],"pa6s4O":["Error while saving ",["0"]],"/JfytP":["Error was ",["err"]],"IBGfrY":["Plugin: ",["0"]],"Se2ost":["Method: ",["0"]],"nDqlBl":"Userinfo","JOUEkZ":["First name: ",["0"]],"GlGzeI":["Last name: ",["0"]],"PkcDO7":["Username: ",["0"]],"M46ISI":"bars","Ai6veK":"oval","8zGXnJ":"dots","gDIqhx":"Design <0/>","QFd2P1":"Primary color","160vo+":"White color","u01284":"Black color","bjp1xg":"Border Radius","EBeoY+":"Loader","xyAcm2":"Found an exsisting login - using it to log you in.","pAtylB":"Not Found","FeQ++0":"Sorry, this page is not known or was moved.","wmCIch":"Go to the start page","OWg6Ht":"Mobile viewport detected","j1oKmM":"Platform UI is optimized for Tablets and Desktops, you can use the official app for a mobile experience.","NtcqDr":"Read the docs","jdbeGb":"login","Phg8Hk":"register","JrBnCJ":["Welcome ",["actionname"]," to"],"hx1ePY":"Placeholder","g87ZP7":"Or continue with email","q8yluz":"Your name","B2Tpo0":"Invalid email","vz42/T":"Password should include at least 6 characters","3xHmj+":"Already have an account? Login","L5ridO":"Don\'t have an account? Register","oFjpQm":"An error occured:"}'
 );
