+++ src/frontend/src/locales/pseudo-LOCALE/messages.ts
@@ -1,3 +1,3 @@
 /*eslint-disable*/ export const messages = JSON.parse(
-  '{"MHrjPM":"Ţĩţĺē","4GKuCs":"Ĺōĝĩń ƒàĩĺēď","jCsNQS":"Ćĥēćķ ŷōũŕ ŷōũŕ ĩńƥũţ àńď ţŕŷ àĝàĩń.","zM9Wd+":"Ĺōĝĩń śũććēśśƒũĺĺ","rxWA39":"Ŵēĺćōmē ƀàćķ!","yfblq9":"Màĩĺ ďēĺĩvēŕŷ śũććēśśƒũĺĺ","R2JMfc":"Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ ţĥē ĺōĝĩń ĺĩńķ. Ĩƒ ŷōũ ĥàvē àń àććōũńţ, ŷōũ ŵĩĺĺ ŕēćēĩvē à ĺōĝĩń ĺĩńķ. Ćĥēćķ ĩń śƥàm ţōō.","ccnxuA":"Ĩńƥũţ ēŕŕōŕ","7sNhEz":"Ũśēŕńàmē","8ZsakT":"Ƥàśśŵōŕď","9TO8nT":"Ŷōũŕ ƥàśśŵōŕď","RfwZxd":"Ŕēśēţ ƥàśśŵōŕď","O3oNi5":"Ēmàĩĺ","Wr5sDQ":"Ŵē ŵĩĺĺ śēńď ŷōũ à ĺĩńķ ţō ĺōĝĩń - ĩƒ ŷōũ àŕē ŕēĝĩśţēŕēď","7ZOmjI":"Śēńď mē àń ēmàĩĺ","XlWstl":"Ĩ ŵĩĺĺ ũśē ũśēŕńàmē àńď ƥàśśŵōŕď","sQia9P":"Ĺōĝ ĩń","F+gz9Z":"Śēńď màĩĺ","Ai2U7L":"Ĥōśţ","6YtxFj":"Ńàmē","yWMzcH":"Ńō ōńē ĥēŕē...","UYWLpE":"Àďď Ĥōśţ","tfDRzk":"Śàvē","IvkbIT":"Ŕēàď Mōŕē","29VNqC":"Ũńķńōŵń ēŕŕōŕ","nlJhkA":"Àń ēŕŕōŕ ōććũŕŕēď:","DVAy0b":"ĨńvēńŢŕēē Ĺōĝō","3TnJRX":"Ţĥĩś ƒēàţũŕē/ƀũţţōń/śĩţē ĩś à ƥĺàćēĥōĺďēŕ ƒōŕ à ƒēàţũŕē ţĥàţ ĩś ńōţ ĩmƥĺēmēńţēď, ōńĺŷ ƥàŕţĩàĺ ōŕ ĩńţēńďēď ƒōŕ ţēśţĩńĝ.","etqXdW":"ƤĹĤ","XDwkfO":"Śćàń ǪŔ ćōďē","GU7xAr":"Ũńķńōŵń ŕēśƥōńśē","UHot+L":"Ēŕŕōŕ ŵĥĩĺē ĝēţţĩńĝ ćàmēŕà","bR26mb":"Ēŕŕōŕ ŵĥĩĺē śćàńńĩńĝ","fvJQqd":"Ēŕŕōŕ ŵĥĩĺē śţōƥƥĩńĝ","CMQ09J":"Śćàńńĩńĝ","Fg9r/3":"Ńōţ śćàńńĩńĝ","QuNKRX":"Śēĺēćţ Ćàmēŕà","m3BKG+":"Śţàŕţ śćàńńĩńĝ","yFRXH8":"Śţōƥ śćàńńĩńĝ","3164SS":"Ńō śćàńś ŷēţ!","RWw9Lg":"Ćĺōśē mōďàĺ","iDNBZe":"Ńōţĩƒĩćàţĩōńś","vERlcd":"Ƥŕōƒĩĺē","Tz0i8g":"Śēţţĩńĝś","r5Xdbs":["Ćũŕŕēńţ ĺàńĝũàĝē ",["locale"]],"XXvCbv":"Śŵĩţćĥ ţō ƥśēũďō ĺàńĝũàĝē","T3FM0r":"Àććōũńţ śēţţĩńĝś","nOhz3x":"Ĺōĝōũţ","rmlxV1":"Ōƥēń Ńàvĩĝàţĩōń","N6Pxr9":"Vĩēŵ àĺĺ","ZDIydz":"Ĝēţ śţàŕţēď","BQDL+H":"Ōvēŕvĩēŵ ōvēŕ ĥĩĝĥ-ĺēvēĺ ōƀĴēćţś, ƒũńćţĩōńś àńď ƥōśśĩƀĺē ũśēćàśēś.","UxKoFf":"Ńàvĩĝàţĩōń","wRR604":"Ƥàĝēś","ohUJJM":"Ƥĺũĝĩńś","TvY/XA":"Ďōćũmēńţàţĩōń","uyJsf6":"Àƀōũţ","hQRttt":"Śũƀmĩţ","dEgA5A":"Ćàńćēĺ","vgP+9p":"Ƥàŕţ","tMMrz4":"ĨńvēńŢŕēē Ďēmō","vu8/DU":"Ĺōćàĺ Śēŕvēŕ","On0aF2":"Ŵēƀśĩţē","RkXlPZ":"ĜĩţĤũƀ","kc+zZA":"Ďēmō","i0qMbr":"Ĥōmē","4GLxhy":"Ĝēţţĩńĝ Śţàŕţēď","VAYCzI":"Ĝēţţĩńĝ śţàŕţēď ŵĩţĥ ĨńvēńŢŕēē","OZtEcz":"ÀƤĨ","aW0h/b":"ĨńvēńŢŕēē ÀƤĨ ďōćũmēńţàţĩōń","BOAupq":"Ďēvēĺōƥēŕ Màńũàĺ","kUcL4g":"ĨńvēńŢŕēē ďēvēĺōƥēŕ màńũàĺ","/lDBHm":"ƑÀǪ","a3pVqb":"Ƒŕēǫũēńţĺŷ àśķēď ǫũēśţĩōńś","kyAi7k":"Ĩńśţàńćē","Q5S3DY":"Àƀōũţ ţĥĩś Ĩńvēńţŕēē ĩńśţàńćē","vHeNia":"ĨńvēńŢŕēē","gfhzPz":"Àƀōũţ ţĥē ĨńvēńŢŕēē ōŕĝ","snyG+w":"Ĺĩćēńśēś","tBjIo1":"Ĺĩćēńśēś ƒōŕ ƥàćķàĝēś ũśēď ƀŷ ĨńvēńŢŕēē","2AZart":"Ōƥēń śōũŕćēà","v+Wp++":"Ōƥēń śōũŕćē","fu2+tK":"Śţàŕţ ƥàĝē ōƒ ŷōũŕ ĩńśţàńćē.","pTE4nz":"Ţĥĩś Ƥōķémōń’ś ćŕŷ ĩś vēŕŷ ĺōũď àńď ďĩśţŕàćţĩńĝ","S+oekQ":"Ţĥĩś Ƥōķémōń’ś ćŕŷ ĩś vēŕŷ ĺōũď àńď ďĩśţŕàćţĩńĝ àńď mōŕē àńď mōŕē àńď mōŕē","kNyJAF":"Ƥŕōƒĩĺē ƥàĝē","CFYxhi":"Ũśēŕ àţţŕĩƀũţēś àńď ďēśĩĝń śēţţĩńĝś.","uP4V6I":"Ƒŕēē ƒōŕ ēvēŕŷōńē","dyMOjI":"Ţĥē ƒĺũĩď ōƒ Śmēàŕĝĺē’ś ţàĩĺ śēćŕēţĩōńś ćĥàńĝēś","d1WpzX":"Ţĥē ƒĺũĩď ōƒ Śmēàŕĝĺē’ś ţàĩĺ śēćŕēţĩōńś ćĥàńĝēś ĩń ţĥē ĩńţēńśĩţŷ","1ekmeV":"àƀć","Wj+wQW":"Ŕàńďōm ĩmàĝē","sGi2sH":"Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ďũĩś àũţēm vēĺ ēũm ĩŕĩũŕē ďōĺōŕ ĩń ĥēńďŕēŕĩţ ĩń vũĺƥũţàţē vēĺĩţ ēśśē mōĺēśţĩē ćōńśēǫũàţ, vēĺ ĩĺĺũm ďōĺōŕē ēũ ƒēũĝĩàţ ńũĺĺà ƒàćĩĺĩśĩś àţ vēŕō ēŕōś ēţ àććũmśàń ēţ ĩũśţō ōďĩō ďĩĝńĩśśĩm ǫũĩ ƀĺàńďĩţ ƥŕàēśēńţ ĺũƥţàţũm źźŕĩĺ ďēĺēńĩţ àũĝũē ďũĩś ďōĺōŕē ţĥē ƒēũĝàĩţ ńũĺĺà ƒàćĩĺĩśĩ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēćţēţũēŕ àďĩƥĩśćĩńĝ ēĺĩţ, śēď ďĩàm ńōńũmmŷ ńĩƀĥ ēũĩśmōď ţĩńćĩďũńţ ũţ ĺàōŕēēţ ďōĺōŕē màĝńà àĺĩǫũàm ēŕàţ vōĺũţƥàţ. Ũţ ŵĩśĩ ēńĩm àď mĩńĩm vēńĩàm, ǫũĩś ńōśţŕũď ēxēŕćĩ ţàţĩōń ũĺĺàmćōŕƥēŕ śũśćĩƥĩţ ĺōƀōŕţĩś ńĩśĺ ũţ àĺĩǫũĩƥ ēx ēà ćōmmōďō ćōńśēǫũàţ. Ďũĩś àũţēm vēĺ ēũm ĩŕĩũŕē ďōĺōŕ ĩń ĥēńďŕēŕĩţ ĩń vũĺƥũţàţē vēĺĩţ ēśśē mōĺēśţĩē ćōńśēǫũàţ, vēĺ ĩĺĺũm ďōĺōŕē ēũ ƒēũĝĩàţ ńũĺĺà ƒàćĩĺĩśĩś àţ vēŕō ēŕōś ēţ àććũmśàń ēţ ĩũśţō ōďĩō ďĩĝńĩśśĩm ǫũĩ ƀĺàńďĩţ ƥŕàēśēńţ ĺũƥţàţũm źźŕĩĺ ďēĺēńĩţ àũĝũē ďũĩś ďōĺōŕē ţĥē ƒēũĝàĩţ ńũĺĺà ƒàćĩĺĩśĩ. Ńàmē ĺĩƀēŕ ţēmƥōŕ ćũm śōĺũţà ńōƀĩś ēĺēĩƒēńď ōƥţĩōń ćōńĝũē ńĩĥĩĺ ĩmƥēŕďĩēţ ďōmĩńĝ ĩď ǫũōď màźĩm ƥĺàćēŕàţ ƒàćēŕ ƥōśśĩm àśśũmē. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēćţēţũēŕ àďĩƥĩśćĩńĝ ēĺĩţ, śēď ďĩàm ńōńũmmŷ ńĩƀĥ ēũĩśmōď ţĩńćĩďũńţ ũţ ĺàōŕēēţ ďōĺōŕē màĝńà àĺĩǫũàm ēŕàţ vōĺũţƥàţ. Ũţ ŵĩśĩ ēńĩm àď mĩńĩm vēńĩàm, ǫũĩś ńōśţŕũď ēxēŕćĩ ţàţĩōń ũĺĺàmćōŕƥēŕ śũśćĩƥĩţ ĺōƀōŕţĩś ńĩśĺ ũţ àĺĩǫũĩƥ ēx ēà ćōmmōďō ćōńśēǫũàţ. Ďũĩś àũţēm vēĺ ēũm ĩŕĩũŕē ďōĺōŕ ĩń ĥēńďŕēŕĩţ ĩń vũĺƥũţàţē vēĺĩţ ēśśē mōĺēśţĩē ćōńśēǫũàţ, vēĺ ĩĺĺũm ďōĺōŕē ēũ ƒēũĝĩàţ ńũĺĺà ƒàćĩĺĩśĩś. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, Àţ àććũśàm àĺĩǫũŷàm ďĩàm ďĩàm ďōĺōŕē ďōĺōŕēś ďũō ēĩŕmōď ēōś ēŕàţ, ēţ ńōńũmŷ śēď ţēmƥōŕ ēţ ēţ ĩńvĩďũńţ Ĵũśţō ĺàƀōŕē Śţēţ ćĺĩţà ēà ēţ ĝũƀēŕĝŕēń, ķàśď màĝńà ńō ŕēƀũm. śàńćţũś śēà śēď ţàķĩmàţà ũţ vēŕō vōĺũƥţũà. ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ. Ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ","K+7Exx":"Ŷàńmà ĩś ćàƥàƀĺē ōƒ śēēĩńĝ 360 ďēĝŕēēś ŵĩţĥōũţ","a3LDKx":"Śēćũŕĩţŷ","D9kxcs":"Ţĥē śĥēĺĺ’ś ŕōũńďēď śĥàƥē àńď ţĥē ĝŕōōvēś ōń ĩţś.","ZlwDi6":"Àńàĺŷţĩćś","Ntb/Ja":"Ţĥĩś Ƥōķémōń ũśēś ĩţś ƒĺŷĩńĝ àƀĩĺĩţŷ ţō ǫũĩćķĺŷ ćĥàśē","q+Lv8f":"Ćōmƀũśķēń ƀàţţĺēś ŵĩţĥ ţĥē ĩńţēńśēĺŷ ĥōţ ƒĺàmēś ĩţ śƥēŵś","FKQcYZ":"Ĺōĝōũţ śũććēśśƒũĺĺ","aJhI/3":"Śēē ŷōũ śōōń.","eX0txO":"Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ à ŕēśēţ ĺĩńķ. Ţĥĩś ōńĺŷ ŵōŕķś ĩƒ ŷōũ ĥàvē àń àććōũńţ. Ćĥēćķ ĩń śƥàm ţōō.","WhimMi":"Ŕēśēţ ƒàĩĺēď","iVj6ge":"Àĺŕēàďŷ ĺōĝĝēď ĩń","FR/+0K":"Ƒōũńď àń ēxĩśţĩńĝ ĺōĝĩń - ũśĩńĝ ĩţ ţō ĺōĝ ŷōũ ĩń.","ps9k8Y":"Ćĥēćķĩńĝ ĩƒ ŷōũ àŕē àĺŕēàďŷ ĺōĝĝēď ĩń","bX1aQ5":"Ńō śēĺēćţĩōń","AA2j+t":"Ēďĩţ ĥōśţ ōƥţĩōńś","eV2FZ+":"Ţōķēń ĩńvàĺĩď","uAHzZQ":"Ŷōũ ńēēď ţō ƥŕōvĩďē à vàĺĩď ţōķēń ţō śēţ à ńēŵ ƥàśśŵōŕď. Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ à ŕēśēţ ĺĩńķ.","+5xxir":"Ńō ţōķēń ƥŕōvĩďēď","KuLTFa":"Ŷōũ ńēēď ţō ƥŕōvĩďē à ţōķēń ţō śēţ à ńēŵ ƥàśśŵōŕď. Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ à ŕēśēţ ĺĩńķ.","Hw2MHB":"Ƥàśśŵōŕď śēţ","+p8fKY":"Ţĥē ƥàśśŵōŕď ŵàś śēţ śũććēśśƒũĺĺŷ. Ŷōũ ćàń ńōŵ ĺōĝĩń ŵĩţĥ ŷōũŕ ńēŵ ƥàśśŵōŕď","V/e7nf":"Śēţ ńēŵ ƥàśśŵōŕď","SlfejT":"Ēŕŕōŕ","TpqeIh":["Ēŕŕōŕ: ",["0"]],"b3ilvM":"Śōŕŕŷ, àń ũńēxƥēćţēď ēŕŕōŕ ĥàś ōććũŕŕēď.","5QTyaY":"Śũƀśćŕĩƀēď Ƥàŕţś","GuGbPw":"Śũƀśćŕĩƀēď Ćàţēĝōŕĩēś","LcKNFQ":"Ĺàţēśţ Ƥàŕţś","eHUZsJ":"ßŌM Ŵàĩţĩńĝ Vàĺĩďàţĩōń","ZopSbj":"Ŕēćēńţĺŷ Ũƥďàţēď","UgdO7s":"Ĺōŵ Śţōćķ","Onj2Pw":"Ďēƥĺēţēď Śţōćķ","Iq/utX":"Ŕēǫũĩŕēď ƒōŕ ßũĩĺď Ōŕďēŕś","ZOsmSm":"Ēxƥĩŕēď Śţōćķ","kc9cAF":"Śţàĺē Śţōćķ","zLhIiS":"ßũĩĺď Ōŕďēŕś Ĩń Ƥŕōĝŕēśś","UBWkDy":"Ōvēŕďũē ßũĩĺď Ōŕďēŕś","WsHr9R":"Ōũţśţàńďĩńĝ Ƥũŕćĥàśē Ōŕďēŕś","fCNzWA":"Ōvēŕďũē Ƥũŕćĥàśē Ōŕďēŕś","gyZThB":"Ōũţśţàńďĩńĝ Śàĺēś Ōŕďēŕś","Gu8K8T":"Ōvēŕďũē Śàĺēś Ōŕďēŕś","XzTq3p":"Ćũŕŕēńţ Ńēŵś","7p5kLi":"Ďàśĥƀōàŕď","edpMcF":"Àũţōũƥďàţē","7PzzBU":"Ũśēŕ","ekfzWq":"Ũśēŕ Śēţţĩńĝś","hFwWnI":"Ńōţĩƒĩćàţĩōń Śēţţĩńĝś","50nnEk":"Ĝĺōƀàĺ Śēţţĩńĝś","eKHY3W":"Ƥĺũĝĩń Śēţţĩńĝś","c6Mp+A":"Śēţţĩńĝś ƒōŕ ţĥē ćũŕŕēńţ ũśēŕ","EBBDLp":"Ĥōmē Ƥàĝē Śēţţĩńĝś","d42r7C":"Śēàŕćĥ Śēţţĩńĝś","o0PqeM":"Ĺàƀēĺ Śēţţĩńĝś","nutMuO":"Ŕēƥōŕţ Śēţţĩńĝś","DdjH42":"Ďĩśƥĺàŷ Śēţţĩńĝś","VzYWwh":"Śēţţĩńĝś ƒōŕ ţĥē ńōţĩƒĩćàţĩōńś","0fzps+":"Ĝĺōƀàĺ Śēŕvēŕ Śēţţĩńĝś","5u2+so":"Ĝĺōƀàĺ Śēţţĩńĝś ƒōŕ ţĥĩś ĩńśţàńćē","S60KP9":"Śēŕvēŕ Śēţţĩńĝś","R+R5Sa":"Ĺōĝĩń Śēţţĩńĝś","NP6Hng":"ßàŕćōďē Śēţţĩńĝś","H2tPtY":"Ƥàŕţ Śēţţĩńĝś","axC9dx":"Ƥŕĩćĩńĝ Śēţţĩńĝś","PN5rCS":"Śţōćķ Śēţţĩńĝś","1PGWAQ":"ßũĩĺď Ōŕďēŕ Śēţţĩńĝś","WObPen":"Ƥũŕćĥàśē Ōŕďēŕ Śēţţĩńĝś","H7F6Gx":"Śàĺēś Ōŕďēŕ Śēţţĩńĝś","Ud411M":"Ƥĺũĝĩń Śēţţĩńĝś ƒōŕ ţĥĩś ĩńśţàńćē","pkdjGY":"Ďàţà ĩś ćũŕŕēńţ ƀēēĩńĝ ĺōàďēď","gIQQwD":"Ƒàĩĺēď ţō ĺōàď","UbtqIw":"Śĥōŵ ĩńţēŕńàĺ ńàmēś","UIwUzc":["Ĩńƥũţ ",["0"]," ĩś ńōţ ķńōŵń"],"J9kB0C":["Śàvēď ćĥàńĝēś ",["0"]],"EOyF2I":["Ćĥàńĝēď ţō ",["0"]],"pa6s4O":["Ēŕŕōŕ ŵĥĩĺē śàvĩńĝ ",["0"]],"/JfytP":["Ēŕŕōŕ ŵàś ",["err"]],"IBGfrY":["Ƥĺũĝĩń: ",["0"]],"Se2ost":["Mēţĥōď: ",["0"]],"nDqlBl":"Ũśēŕĩńƒō","JOUEkZ":["Ƒĩŕśţ ńàmē: ",["0"]],"GlGzeI":["Ĺàśţ ńàmē: ",["0"]],"PkcDO7":["Ũśēŕńàmē: ",["0"]],"xyAcm2":"Ƒōũńď àń ēxśĩśţĩńĝ ĺōĝĩń - ũśĩńĝ ĩţ ţō ĺōĝ ŷōũ ĩń.","pAtylB":"Ńōţ Ƒōũńď","FeQ++0":"Śōŕŕŷ, ţĥĩś ƥàĝē ĩś ńōţ ķńōŵń ōŕ ŵàś mōvēď.","wmCIch":"Ĝō ţō ţĥē śţàŕţ ƥàĝē","OWg6Ht":"Mōƀĩĺē vĩēŵƥōŕţ ďēţēćţēď","j1oKmM":"Ƥĺàţƒōŕm ŨĨ ĩś ōƥţĩmĩźēď ƒōŕ Ţàƀĺēţś àńď Ďēśķţōƥś, ŷōũ ćàń ũśē ţĥē ōƒƒĩćĩàĺ àƥƥ ƒōŕ à mōƀĩĺē ēxƥēŕĩēńćē.","NtcqDr":"Ŕēàď ţĥē ďōćś","jdbeGb":"ĺōĝĩń","Phg8Hk":"ŕēĝĩśţēŕ","JrBnCJ":["Ŵēĺćōmē ",["actionname"]," ţō"],"hx1ePY":"Ƥĺàćēĥōĺďēŕ","g87ZP7":"Ōŕ ćōńţĩńũē ŵĩţĥ ēmàĩĺ","q8yluz":"Ŷōũŕ ńàmē","B2Tpo0":"Ĩńvàĺĩď ēmàĩĺ","vz42/T":"Ƥàśśŵōŕď śĥōũĺď ĩńćĺũďē àţ ĺēàśţ 6 ćĥàŕàćţēŕś","3xHmj+":"Àĺŕēàďŷ ĥàvē àń àććōũńţ? Ĺōĝĩń","L5ridO":"Ďōń\'ţ ĥàvē àń àććōũńţ? Ŕēĝĩśţēŕ","oFjpQm":"Àń ēŕŕōŕ ōććũŕēď:"}'
+  '{"MHrjPM":"Ţĩţĺē","4GKuCs":"Ĺōĝĩń ƒàĩĺēď","jCsNQS":"Ćĥēćķ ŷōũŕ ŷōũŕ ĩńƥũţ àńď ţŕŷ àĝàĩń.","zM9Wd+":"Ĺōĝĩń śũććēśśƒũĺĺ","rxWA39":"Ŵēĺćōmē ƀàćķ!","yfblq9":"Màĩĺ ďēĺĩvēŕŷ śũććēśśƒũĺĺ","R2JMfc":"Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ ţĥē ĺōĝĩń ĺĩńķ. Ĩƒ ŷōũ ĥàvē àń àććōũńţ, ŷōũ ŵĩĺĺ ŕēćēĩvē à ĺōĝĩń ĺĩńķ. Ćĥēćķ ĩń śƥàm ţōō.","ccnxuA":"Ĩńƥũţ ēŕŕōŕ","7sNhEz":"Ũśēŕńàmē","8ZsakT":"Ƥàśśŵōŕď","9TO8nT":"Ŷōũŕ ƥàśśŵōŕď","RfwZxd":"Ŕēśēţ ƥàśśŵōŕď","O3oNi5":"Ēmàĩĺ","Wr5sDQ":"Ŵē ŵĩĺĺ śēńď ŷōũ à ĺĩńķ ţō ĺōĝĩń - ĩƒ ŷōũ àŕē ŕēĝĩśţēŕēď","7ZOmjI":"Śēńď mē àń ēmàĩĺ","XlWstl":"Ĩ ŵĩĺĺ ũśē ũśēŕńàmē àńď ƥàśśŵōŕď","sQia9P":"Ĺōĝ ĩń","F+gz9Z":"Śēńď màĩĺ","Ai2U7L":"Ĥōśţ","6YtxFj":"Ńàmē","yWMzcH":"Ńō ōńē ĥēŕē...","UYWLpE":"Àďď Ĥōśţ","tfDRzk":"Śàvē","IvkbIT":"Ŕēàď Mōŕē","29VNqC":"Ũńķńōŵń ēŕŕōŕ","nlJhkA":"Àń ēŕŕōŕ ōććũŕŕēď:","DVAy0b":"ĨńvēńŢŕēē Ĺōĝō","3TnJRX":"Ţĥĩś ƒēàţũŕē/ƀũţţōń/śĩţē ĩś à ƥĺàćēĥōĺďēŕ ƒōŕ à ƒēàţũŕē ţĥàţ ĩś ńōţ ĩmƥĺēmēńţēď, ōńĺŷ ƥàŕţĩàĺ ōŕ ĩńţēńďēď ƒōŕ ţēśţĩńĝ.","etqXdW":"ƤĹĤ","XDwkfO":"Śćàń ǪŔ ćōďē","GU7xAr":"Ũńķńōŵń ŕēśƥōńśē","UHot+L":"Ēŕŕōŕ ŵĥĩĺē ĝēţţĩńĝ ćàmēŕà","bR26mb":"Ēŕŕōŕ ŵĥĩĺē śćàńńĩńĝ","fvJQqd":"Ēŕŕōŕ ŵĥĩĺē śţōƥƥĩńĝ","CMQ09J":"Śćàńńĩńĝ","Fg9r/3":"Ńōţ śćàńńĩńĝ","QuNKRX":"Śēĺēćţ Ćàmēŕà","m3BKG+":"Śţàŕţ śćàńńĩńĝ","yFRXH8":"Śţōƥ śćàńńĩńĝ","3164SS":"Ńō śćàńś ŷēţ!","RWw9Lg":"Ćĺōśē mōďàĺ","iDNBZe":"Ńōţĩƒĩćàţĩōńś","vERlcd":"Ƥŕōƒĩĺē","Tz0i8g":"Śēţţĩńĝś","r5Xdbs":["Ćũŕŕēńţ ĺàńĝũàĝē ",["locale"]],"XXvCbv":"Śŵĩţćĥ ţō ƥśēũďō ĺàńĝũàĝē","T3FM0r":"Àććōũńţ śēţţĩńĝś","nOhz3x":"Ĺōĝōũţ","rmlxV1":"Ōƥēń Ńàvĩĝàţĩōń","N6Pxr9":"Vĩēŵ àĺĺ","ZDIydz":"Ĝēţ śţàŕţēď","BQDL+H":"Ōvēŕvĩēŵ ōvēŕ ĥĩĝĥ-ĺēvēĺ ōƀĴēćţś, ƒũńćţĩōńś àńď ƥōśśĩƀĺē ũśēćàśēś.","UxKoFf":"Ńàvĩĝàţĩōń","wRR604":"Ƥàĝēś","ohUJJM":"Ƥĺũĝĩńś","TvY/XA":"Ďōćũmēńţàţĩōń","uyJsf6":"Àƀōũţ","hQRttt":"Śũƀmĩţ","dEgA5A":"Ćàńćēĺ","vgP+9p":"Ƥàŕţ","tMMrz4":"ĨńvēńŢŕēē Ďēmō","vu8/DU":"Ĺōćàĺ Śēŕvēŕ","On0aF2":"Ŵēƀśĩţē","RkXlPZ":"ĜĩţĤũƀ","kc+zZA":"Ďēmō","i0qMbr":"Ĥōmē","4GLxhy":"Ĝēţţĩńĝ Śţàŕţēď","VAYCzI":"Ĝēţţĩńĝ śţàŕţēď ŵĩţĥ ĨńvēńŢŕēē","OZtEcz":"ÀƤĨ","aW0h/b":"ĨńvēńŢŕēē ÀƤĨ ďōćũmēńţàţĩōń","BOAupq":"Ďēvēĺōƥēŕ Màńũàĺ","kUcL4g":"ĨńvēńŢŕēē ďēvēĺōƥēŕ màńũàĺ","/lDBHm":"ƑÀǪ","a3pVqb":"Ƒŕēǫũēńţĺŷ àśķēď ǫũēśţĩōńś","kyAi7k":"Ĩńśţàńćē","Q5S3DY":"Àƀōũţ ţĥĩś Ĩńvēńţŕēē ĩńśţàńćē","vHeNia":"ĨńvēńŢŕēē","gfhzPz":"Àƀōũţ ţĥē ĨńvēńŢŕēē ōŕĝ","snyG+w":"Ĺĩćēńśēś","tBjIo1":"Ĺĩćēńśēś ƒōŕ ƥàćķàĝēś ũśēď ƀŷ ĨńvēńŢŕēē","2AZart":"Ōƥēń śōũŕćēà","v+Wp++":"Ōƥēń śōũŕćē","fu2+tK":"Śţàŕţ ƥàĝē ōƒ ŷōũŕ ĩńśţàńćē.","pTE4nz":"Ţĥĩś Ƥōķémōń’ś ćŕŷ ĩś vēŕŷ ĺōũď àńď ďĩśţŕàćţĩńĝ","S+oekQ":"Ţĥĩś Ƥōķémōń’ś ćŕŷ ĩś vēŕŷ ĺōũď àńď ďĩśţŕàćţĩńĝ àńď mōŕē àńď mōŕē àńď mōŕē","kNyJAF":"Ƥŕōƒĩĺē ƥàĝē","CFYxhi":"Ũśēŕ àţţŕĩƀũţēś àńď ďēśĩĝń śēţţĩńĝś.","uP4V6I":"Ƒŕēē ƒōŕ ēvēŕŷōńē","dyMOjI":"Ţĥē ƒĺũĩď ōƒ Śmēàŕĝĺē’ś ţàĩĺ śēćŕēţĩōńś ćĥàńĝēś","d1WpzX":"Ţĥē ƒĺũĩď ōƒ Śmēàŕĝĺē’ś ţàĩĺ śēćŕēţĩōńś ćĥàńĝēś ĩń ţĥē ĩńţēńśĩţŷ","1ekmeV":"àƀć","Wj+wQW":"Ŕàńďōm ĩmàĝē","sGi2sH":"Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ďũĩś àũţēm vēĺ ēũm ĩŕĩũŕē ďōĺōŕ ĩń ĥēńďŕēŕĩţ ĩń vũĺƥũţàţē vēĺĩţ ēśśē mōĺēśţĩē ćōńśēǫũàţ, vēĺ ĩĺĺũm ďōĺōŕē ēũ ƒēũĝĩàţ ńũĺĺà ƒàćĩĺĩśĩś àţ vēŕō ēŕōś ēţ àććũmśàń ēţ ĩũśţō ōďĩō ďĩĝńĩśśĩm ǫũĩ ƀĺàńďĩţ ƥŕàēśēńţ ĺũƥţàţũm źźŕĩĺ ďēĺēńĩţ àũĝũē ďũĩś ďōĺōŕē ţĥē ƒēũĝàĩţ ńũĺĺà ƒàćĩĺĩśĩ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēćţēţũēŕ àďĩƥĩśćĩńĝ ēĺĩţ, śēď ďĩàm ńōńũmmŷ ńĩƀĥ ēũĩśmōď ţĩńćĩďũńţ ũţ ĺàōŕēēţ ďōĺōŕē màĝńà àĺĩǫũàm ēŕàţ vōĺũţƥàţ. Ũţ ŵĩśĩ ēńĩm àď mĩńĩm vēńĩàm, ǫũĩś ńōśţŕũď ēxēŕćĩ ţàţĩōń ũĺĺàmćōŕƥēŕ śũśćĩƥĩţ ĺōƀōŕţĩś ńĩśĺ ũţ àĺĩǫũĩƥ ēx ēà ćōmmōďō ćōńśēǫũàţ. Ďũĩś àũţēm vēĺ ēũm ĩŕĩũŕē ďōĺōŕ ĩń ĥēńďŕēŕĩţ ĩń vũĺƥũţàţē vēĺĩţ ēśśē mōĺēśţĩē ćōńśēǫũàţ, vēĺ ĩĺĺũm ďōĺōŕē ēũ ƒēũĝĩàţ ńũĺĺà ƒàćĩĺĩśĩś àţ vēŕō ēŕōś ēţ àććũmśàń ēţ ĩũśţō ōďĩō ďĩĝńĩśśĩm ǫũĩ ƀĺàńďĩţ ƥŕàēśēńţ ĺũƥţàţũm źźŕĩĺ ďēĺēńĩţ àũĝũē ďũĩś ďōĺōŕē ţĥē ƒēũĝàĩţ ńũĺĺà ƒàćĩĺĩśĩ. Ńàmē ĺĩƀēŕ ţēmƥōŕ ćũm śōĺũţà ńōƀĩś ēĺēĩƒēńď ōƥţĩōń ćōńĝũē ńĩĥĩĺ ĩmƥēŕďĩēţ ďōmĩńĝ ĩď ǫũōď màźĩm ƥĺàćēŕàţ ƒàćēŕ ƥōśśĩm àśśũmē. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēćţēţũēŕ àďĩƥĩśćĩńĝ ēĺĩţ, śēď ďĩàm ńōńũmmŷ ńĩƀĥ ēũĩśmōď ţĩńćĩďũńţ ũţ ĺàōŕēēţ ďōĺōŕē màĝńà àĺĩǫũàm ēŕàţ vōĺũţƥàţ. Ũţ ŵĩśĩ ēńĩm àď mĩńĩm vēńĩàm, ǫũĩś ńōśţŕũď ēxēŕćĩ ţàţĩōń ũĺĺàmćōŕƥēŕ śũśćĩƥĩţ ĺōƀōŕţĩś ńĩśĺ ũţ àĺĩǫũĩƥ ēx ēà ćōmmōďō ćōńśēǫũàţ. Ďũĩś àũţēm vēĺ ēũm ĩŕĩũŕē ďōĺōŕ ĩń ĥēńďŕēŕĩţ ĩń vũĺƥũţàţē vēĺĩţ ēśśē mōĺēśţĩē ćōńśēǫũàţ, vēĺ ĩĺĺũm ďōĺōŕē ēũ ƒēũĝĩàţ ńũĺĺà ƒàćĩĺĩśĩś. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, Àţ àććũśàm àĺĩǫũŷàm ďĩàm ďĩàm ďōĺōŕē ďōĺōŕēś ďũō ēĩŕmōď ēōś ēŕàţ, ēţ ńōńũmŷ śēď ţēmƥōŕ ēţ ēţ ĩńvĩďũńţ Ĵũśţō ĺàƀōŕē Śţēţ ćĺĩţà ēà ēţ ĝũƀēŕĝŕēń, ķàśď màĝńà ńō ŕēƀũm. śàńćţũś śēà śēď ţàķĩmàţà ũţ vēŕō vōĺũƥţũà. ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ. Ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ, ćōńśēţēţũŕ śàďĩƥśćĩńĝ ēĺĩţŕ, śēď ďĩàm ńōńũmŷ ēĩŕmōď ţēmƥōŕ ĩńvĩďũńţ ũţ ĺàƀōŕē ēţ ďōĺōŕē màĝńà àĺĩǫũŷàm ēŕàţ, śēď ďĩàm vōĺũƥţũà. Àţ vēŕō ēōś ēţ àććũśàm ēţ Ĵũśţō ďũō ďōĺōŕēś ēţ ēà ŕēƀũm. Śţēţ ćĺĩţà ķàśď ĝũƀēŕĝŕēń, ńō śēà ţàķĩmàţà śàńćţũś ēśţ Ĺōŕēm ĩƥśũm ďōĺōŕ śĩţ àmēţ. Ĺōŕēm ĩƥśũm ďōĺōŕ","K+7Exx":"Ŷàńmà ĩś ćàƥàƀĺē ōƒ śēēĩńĝ 360 ďēĝŕēēś ŵĩţĥōũţ","a3LDKx":"Śēćũŕĩţŷ","D9kxcs":"Ţĥē śĥēĺĺ’ś ŕōũńďēď śĥàƥē àńď ţĥē ĝŕōōvēś ōń ĩţś.","ZlwDi6":"Àńàĺŷţĩćś","Ntb/Ja":"Ţĥĩś Ƥōķémōń ũśēś ĩţś ƒĺŷĩńĝ àƀĩĺĩţŷ ţō ǫũĩćķĺŷ ćĥàśē","q+Lv8f":"Ćōmƀũśķēń ƀàţţĺēś ŵĩţĥ ţĥē ĩńţēńśēĺŷ ĥōţ ƒĺàmēś ĩţ śƥēŵś","FKQcYZ":"Ĺōĝōũţ śũććēśśƒũĺĺ","aJhI/3":"Śēē ŷōũ śōōń.","eX0txO":"Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ à ŕēśēţ ĺĩńķ. Ţĥĩś ōńĺŷ ŵōŕķś ĩƒ ŷōũ ĥàvē àń àććōũńţ. Ćĥēćķ ĩń śƥàm ţōō.","WhimMi":"Ŕēśēţ ƒàĩĺēď","iVj6ge":"Àĺŕēàďŷ ĺōĝĝēď ĩń","FR/+0K":"Ƒōũńď àń ēxĩśţĩńĝ ĺōĝĩń - ũśĩńĝ ĩţ ţō ĺōĝ ŷōũ ĩń.","ps9k8Y":"Ćĥēćķĩńĝ ĩƒ ŷōũ àŕē àĺŕēàďŷ ĺōĝĝēď ĩń","bX1aQ5":"Ńō śēĺēćţĩōń","AA2j+t":"Ēďĩţ ĥōśţ ōƥţĩōńś","eV2FZ+":"Ţōķēń ĩńvàĺĩď","uAHzZQ":"Ŷōũ ńēēď ţō ƥŕōvĩďē à vàĺĩď ţōķēń ţō śēţ à ńēŵ ƥàśśŵōŕď. Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ à ŕēśēţ ĺĩńķ.","+5xxir":"Ńō ţōķēń ƥŕōvĩďēď","KuLTFa":"Ŷōũ ńēēď ţō ƥŕōvĩďē à ţōķēń ţō śēţ à ńēŵ ƥàśśŵōŕď. Ćĥēćķ ŷōũŕ ĩńƀōx ƒōŕ à ŕēśēţ ĺĩńķ.","Hw2MHB":"Ƥàśśŵōŕď śēţ","+p8fKY":"Ţĥē ƥàśśŵōŕď ŵàś śēţ śũććēśśƒũĺĺŷ. Ŷōũ ćàń ńōŵ ĺōĝĩń ŵĩţĥ ŷōũŕ ńēŵ ƥàśśŵōŕď","V/e7nf":"Śēţ ńēŵ ƥàśśŵōŕď","SlfejT":"Ēŕŕōŕ","TpqeIh":["Ēŕŕōŕ: ",["0"]],"b3ilvM":"Śōŕŕŷ, àń ũńēxƥēćţēď ēŕŕōŕ ĥàś ōććũŕŕēď.","5QTyaY":"Śũƀśćŕĩƀēď Ƥàŕţś","GuGbPw":"Śũƀśćŕĩƀēď Ćàţēĝōŕĩēś","LcKNFQ":"Ĺàţēśţ Ƥàŕţś","eHUZsJ":"ßŌM Ŵàĩţĩńĝ Vàĺĩďàţĩōń","ZopSbj":"Ŕēćēńţĺŷ Ũƥďàţēď","UgdO7s":"Ĺōŵ Śţōćķ","Onj2Pw":"Ďēƥĺēţēď Śţōćķ","Iq/utX":"Ŕēǫũĩŕēď ƒōŕ ßũĩĺď Ōŕďēŕś","ZOsmSm":"Ēxƥĩŕēď Śţōćķ","kc9cAF":"Śţàĺē Śţōćķ","zLhIiS":"ßũĩĺď Ōŕďēŕś Ĩń Ƥŕōĝŕēśś","UBWkDy":"Ōvēŕďũē ßũĩĺď Ōŕďēŕś","WsHr9R":"Ōũţśţàńďĩńĝ Ƥũŕćĥàśē Ōŕďēŕś","fCNzWA":"Ōvēŕďũē Ƥũŕćĥàśē Ōŕďēŕś","gyZThB":"Ōũţśţàńďĩńĝ Śàĺēś Ōŕďēŕś","Gu8K8T":"Ōvēŕďũē Śàĺēś Ōŕďēŕś","XzTq3p":"Ćũŕŕēńţ Ńēŵś","7p5kLi":"Ďàśĥƀōàŕď","edpMcF":"Àũţōũƥďàţē","7PzzBU":"Ũśēŕ","ekfzWq":"Ũśēŕ Śēţţĩńĝś","hFwWnI":"Ńōţĩƒĩćàţĩōń Śēţţĩńĝś","50nnEk":"Ĝĺōƀàĺ Śēţţĩńĝś","eKHY3W":"Ƥĺũĝĩń Śēţţĩńĝś","c6Mp+A":"Śēţţĩńĝś ƒōŕ ţĥē ćũŕŕēńţ ũśēŕ","EBBDLp":"Ĥōmē Ƥàĝē Śēţţĩńĝś","d42r7C":"Śēàŕćĥ Śēţţĩńĝś","o0PqeM":"Ĺàƀēĺ Śēţţĩńĝś","nutMuO":"Ŕēƥōŕţ Śēţţĩńĝś","DdjH42":"Ďĩśƥĺàŷ Śēţţĩńĝś","VzYWwh":"Śēţţĩńĝś ƒōŕ ţĥē ńōţĩƒĩćàţĩōńś","0fzps+":"Ĝĺōƀàĺ Śēŕvēŕ Śēţţĩńĝś","5u2+so":"Ĝĺōƀàĺ Śēţţĩńĝś ƒōŕ ţĥĩś ĩńśţàńćē","S60KP9":"Śēŕvēŕ Śēţţĩńĝś","R+R5Sa":"Ĺōĝĩń Śēţţĩńĝś","NP6Hng":"ßàŕćōďē Śēţţĩńĝś","H2tPtY":"Ƥàŕţ Śēţţĩńĝś","axC9dx":"Ƥŕĩćĩńĝ Śēţţĩńĝś","PN5rCS":"Śţōćķ Śēţţĩńĝś","1PGWAQ":"ßũĩĺď Ōŕďēŕ Śēţţĩńĝś","WObPen":"Ƥũŕćĥàśē Ōŕďēŕ Śēţţĩńĝś","H7F6Gx":"Śàĺēś Ōŕďēŕ Śēţţĩńĝś","Ud411M":"Ƥĺũĝĩń Śēţţĩńĝś ƒōŕ ţĥĩś ĩńśţàńćē","pkdjGY":"Ďàţà ĩś ćũŕŕēńţ ƀēēĩńĝ ĺōàďēď","gIQQwD":"Ƒàĩĺēď ţō ĺōàď","UbtqIw":"Śĥōŵ ĩńţēŕńàĺ ńàmēś","UIwUzc":["Ĩńƥũţ ",["0"]," ĩś ńōţ ķńōŵń"],"J9kB0C":["Śàvēď ćĥàńĝēś ",["0"]],"EOyF2I":["Ćĥàńĝēď ţō ",["0"]],"pa6s4O":["Ēŕŕōŕ ŵĥĩĺē śàvĩńĝ ",["0"]],"/JfytP":["Ēŕŕōŕ ŵàś ",["err"]],"IBGfrY":["Ƥĺũĝĩń: ",["0"]],"Se2ost":["Mēţĥōď: ",["0"]],"nDqlBl":"Ũśēŕĩńƒō","JOUEkZ":["Ƒĩŕśţ ńàmē: ",["0"]],"GlGzeI":["Ĺàśţ ńàmē: ",["0"]],"PkcDO7":["Ũśēŕńàmē: ",["0"]],"M46ISI":"ƀàŕś","Ai6veK":"ōvàĺ","8zGXnJ":"ďōţś","gDIqhx":"Ďēśĩĝń <0/>","QFd2P1":"Ƥŕĩmàŕŷ ćōĺōŕ","160vo+":"Ŵĥĩţē ćōĺōŕ","u01284":"ßĺàćķ ćōĺōŕ","bjp1xg":"ßōŕďēŕ Ŕàďĩũś","EBeoY+":"Ĺōàďēŕ","xyAcm2":"Ƒōũńď àń ēxśĩśţĩńĝ ĺōĝĩń - ũśĩńĝ ĩţ ţō ĺōĝ ŷōũ ĩń.","pAtylB":"Ńōţ Ƒōũńď","FeQ++0":"Śōŕŕŷ, ţĥĩś ƥàĝē ĩś ńōţ ķńōŵń ōŕ ŵàś mōvēď.","wmCIch":"Ĝō ţō ţĥē śţàŕţ ƥàĝē","OWg6Ht":"Mōƀĩĺē vĩēŵƥōŕţ ďēţēćţēď","j1oKmM":"Ƥĺàţƒōŕm ŨĨ ĩś ōƥţĩmĩźēď ƒōŕ Ţàƀĺēţś àńď Ďēśķţōƥś, ŷōũ ćàń ũśē ţĥē ōƒƒĩćĩàĺ àƥƥ ƒōŕ à mōƀĩĺē ēxƥēŕĩēńćē.","NtcqDr":"Ŕēàď ţĥē ďōćś","jdbeGb":"ĺōĝĩń","Phg8Hk":"ŕēĝĩśţēŕ","JrBnCJ":["Ŵēĺćōmē ",["actionname"]," ţō"],"hx1ePY":"Ƥĺàćēĥōĺďēŕ","g87ZP7":"Ōŕ ćōńţĩńũē ŵĩţĥ ēmàĩĺ","q8yluz":"Ŷōũŕ ńàmē","B2Tpo0":"Ĩńvàĺĩď ēmàĩĺ","vz42/T":"Ƥàśśŵōŕď śĥōũĺď ĩńćĺũďē àţ ĺēàśţ 6 ćĥàŕàćţēŕś","3xHmj+":"Àĺŕēàďŷ ĥàvē àń àććōũńţ? Ĺōĝĩń","L5ridO":"Ďōń\'ţ ĥàvē àń àććōũńţ? Ŕēĝĩśţēŕ","oFjpQm":"Àń ēŕŕōŕ ōććũŕēď:"}'
 );
