+++ src/frontend/src/pages/Index/Profile/UserPanel.tsx
@@ -1,33 +1,22 @@
 import { Trans } from '@lingui/macro';
 import {
   Button,
-  ColorInput,
-  ColorPicker,
   Container,
-  DEFAULT_THEME,
   Grid,
   Group,
-  Select,
   SimpleGrid,
   Skeleton,
-  Slider,
-  Table,
   Text,
   TextInput,
   Title
 } from '@mantine/core';
 import { useForm } from '@mantine/form';
 import { useToggle } from '@mantine/hooks';
-import { LoaderType } from '@mantine/styles/lib/theme/types/MantineTheme';
 import { useQuery } from '@tanstack/react-query';
-import { useState } from 'react';
 
 import { api, queryClient } from '../../../App';
 import { EditButton } from '../../../components/items/EditButton';
-import { PlaceholderPill } from '../../../components/items/Placeholder';
-import { SizeMarks } from '../../../defaults/defaults';
-import { InvenTreeStyle } from '../../../globalStyle';
-import { useLocalState } from '../../../states/LocalState';
+import { UserTheme } from './UserTheme';
 
 export function UserPanel() {
   // view
@@ -44,7 +33,7 @@ export function UserPanel() {
   });
 
   return (
-    <Container>
+    <div>
       <SimpleGrid cols={2} spacing="md">
         <Container w="100%">
           {isLoading ? (
@@ -65,7 +54,7 @@ export function UserPanel() {
           </Grid.Col>
         </Grid>
       </SimpleGrid>
-    </Container>
+    </div>
   );
 }
 
@@ -134,117 +123,5 @@ export function UserInfo({ data }: { data: any }) {
         </Group>
       ) : null}
     </form>
-  );
-}
-
-export function UserTheme({ height }: { height: number }) {
-  const { theme } = InvenTreeStyle();
-
-  function getLkp(color: string) {
-    return { [DEFAULT_THEME.colors[color][6]]: color };
-  }
-  const lookup = Object.assign(
-    {},
-    ...Object.keys(DEFAULT_THEME.colors).map((clr) => getLkp(clr))
-  );
-
-  // primary color
-  function changePrimary(color: string) {
-    useLocalState.setState({ primaryColor: lookup[color] });
-  }
-  // white color
-  const [whiteColor, setWhiteColor] = useState(theme.white);
-  function changeWhite(color: string) {
-    useLocalState.setState({ whiteColor: color });
-    setWhiteColor(color);
-  }
-  // black color
-  const [blackColor, setBlackColor] = useState(theme.black);
-  function changeBlack(color: string) {
-    useLocalState.setState({ blackColor: color });
-    setBlackColor(color);
-  }
-  // radius
-  function getMark(value: number) {
-    const obj = SizeMarks.find((mark) => mark.value === value);
-    if (obj) return obj;
-    return SizeMarks[0];
-  }
-  function getDefaultRadius() {
-    const obj = SizeMarks.find(
-      (mark) => mark.label === useLocalState.getState().radius
-    );
-    if (obj) return obj.value;
-    return 50;
-  }
-  const [radius, setRadius] = useState(getDefaultRadius());
-  function changeRadius(value: number) {
-    setRadius(value);
-    useLocalState.setState({ radius: getMark(value).label });
-  }
-  // loader
-  const loaderDate = ['bars', 'oval', 'dots'];
-  const [loader, setLoader] = useState<LoaderType>(theme.loader);
-  function changeLoader(value: LoaderType) {
-    setLoader(value);
-    useLocalState.setState({ loader: value });
-  }
-
-  return (
-    <Container w="100%" mih={height}>
-      <Title order={3}>
-        Design <PlaceholderPill />
-      </Title>
-      <Table>
-        <tbody>
-          <tr>
-            <td>Primary color</td>
-            <td>
-              <ColorPicker
-                format="hex"
-                onChange={changePrimary}
-                withPicker={false}
-                swatches={Object.keys(lookup)}
-              />
-            </td>
-          </tr>
-          <tr>
-            <td>White color</td>
-            <td>
-              <ColorInput value={whiteColor} onChange={changeWhite} />
-            </td>
-          </tr>
-          <tr>
-            <td>Black color</td>
-            <td>
-              <ColorInput value={blackColor} onChange={changeBlack} />
-            </td>
-          </tr>
-          <tr>
-            <td>Border Radius</td>
-            <td>
-              <Slider
-                label={(val) => getMark(val).label}
-                defaultValue={50}
-                step={25}
-                marks={SizeMarks}
-                value={radius}
-                onChange={changeRadius}
-              />
-            </td>
-          </tr>
-          <tr>
-            <td>Loader</td>
-            <td>
-              <Select
-                data={loaderDate}
-                value={loader}
-                onChange={changeLoader}
-              />
-            </td>
-          </tr>
-        </tbody>
-      </Table>
-    </Container>
   );
 }
